+++ NEWS.md
@@ -1,5 +1,6 @@
 # esquisse 1.2.0
 
+* Warning messages are now displayed only once per session, and the user can choose via the settings menu not to display them at all. If you're using the `esquisse_server()` module, there's a new `notify_warnings =` argument for choosing what to do with notifications: display them all, only once per warning or never display them at all.
 * `esquisse_server()` : support for reactive functions for `data_rv` argument.
 * Allow `esquisse_server()` to be initialized with NULL without forcing the import module to appear using `import_from = NULL` [#232](https://github.com/dreamRs/esquisse/issues/232).
 * New geom available in the interface : `geom_path`.